+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version 10.1.0.1
 
-* Added support for Merchants, Loot, Character, Equipment Flyouts
+* Added support for Merchants, Loot, Character, Inspect Character, Equipment Flyouts
 * Added support for Equipment Manager (Wrath Classic only)
 * Added support for Bag Bar (Wrath Classic and Classic Era only)